wellbeing: Fix some unsigned comparisons against zero
These highlighted a bug in the types here: `offset`/`model_offset` were
signed, but `i` was unsigned. The addition of the two was implicitly
promoted to unsigned, but the intent of the code was for it to be signed
(and negative results caught).

Fix that by changing `i` to be signed.

Fixes warnings:
```
../panels/wellbeing/cc-screen-time-statistics-row.c: In function ‘calculate_total_screen_time_for_week’:
../panels/wellbeing/cc-screen-time-statistics-row.c:617:24: error: comparison of unsigned expression in ‘>= 0’ is always true [-Werror=type-limits]
  617 |     sum += (offset + i >= 0 && offset + i < self->model.n_days) ? self->model.screen_time_per_day[offset + i] : 0;
      |                        ^~
../panels/wellbeing/cc-screen-time-statistics-row.c: In function ‘update_ui_for_model_or_selected_date’:
../panels/wellbeing/cc-screen-time-statistics-row.c:963:39: error: comparison of unsigned expression in ‘>= 0’ is always true [-Werror=type-limits]
  963 |     data_slice[i] = (model_offset + i >= 0 && model_offset + i < self->model.n_days) ? self->model.screen_time_per_day[model_offset + i] : NAN;
      |                                       ^~
```

+++ panels/wellbeing/cc-screen-time-statistics-row.c
@@ -613,7 +613,7 @@ calculate_total_screen_time_for_week (CcScreenTimeStatisticsRow *self,
   const int offset = g_date_days_between (&self->model.start_date, first_day_of_week);
 
   sum = 0.0;
-  for (size_t i = 0; i < 7; i++)
+  for (int i = 0; i < 7; i++)
     sum += (offset + i >= 0 && offset + i < self->model.n_days) ? self->model.screen_time_per_day[offset + i] : 0;
 
   return sum;
@@ -959,7 +959,7 @@ update_ui_for_model_or_selected_date (CcScreenTimeStatisticsRow *self)
    * buffer to avoid that, and initialise its values to NAN to indicate
    * unknown/unset data values. */
   data_slice = g_new (double, 7);
-  for (size_t i = 0; i < 7; i++)
+  for (int i = 0; i < 7; i++)
     data_slice[i] = (model_offset + i >= 0 && model_offset + i < self->model.n_days) ? self->model.screen_time_per_day[model_offset + i] : NAN;
 
   cc_bar_chart_set_data (self->bar_chart, data_slice, 7);
